[📂 file/#47]: InternshipDetailService 인터페이스화
- InternshipDetailService와 InternshipDetailServiceImpl로 분리
